requirements.txt
pillow
pickle
streamlit
numpy
keras
tensorflow
os
